+++ src/ts/Core/Models/ABGroup.ts
@@ -90,10 +90,11 @@ export const FakeZyngaFilteredReverseABTest = new ZyngaFilteredReverseABTest(16,
 export const MabReverseABTest = new ZyngaFilteredReverseABTest(7);
 export const GooglePlayDetectionTest = new ABTest(8);
 export const LoadV5NoInvalidation = new ABTest(13);
-export const LoadV5 = new ReverseABTest(14, 15, 18, 19);
+export const LoadV5 = new ABTest(14, 15, 18, 19);
 
 // Disabled AB Tests
 export const OpenMeasurementTest = new DisabledABTest();
 export const IframeEndcardTest = new DisabledABTest();
 export const ConsentUXTest = new DisabledABTest();
 export const PrivacySDKTest = new DisabledABTest();
+export const MraidWebplayerTest = new DisabledABTest();